+++ src/common/DetailsActions/DetailsExportAction/DetailsExportAction.test.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import user from '@testing-library/user-event';
+
+import { DetailsExportAction } from './DetailsExportAction';
+
+const defaultProps = {
+  onExportCSV: jest.fn(),
+  perm: 'perm',
+  toggleActionMenu: jest.fn(),
+};
+
+const renderDetailsExportAction = (props = {}) => render(
+  <DetailsExportAction
+    {...defaultProps}
+    {...props}
+  />,
+);
+
+describe('DetailsExportAction', () => {
+  it('should display action button', () => {
+    renderDetailsExportAction();
+
+    screen.debug();
+
+    expect(screen.getByTestId('action-export-csv')).toBeInTheDocument();
+  });
+
+  it('should call \'onExportCSV\' when the button was clicked', () => {
+    renderDetailsExportAction();
+
+    user.click(screen.getByTestId('action-export-csv'));
+
+    expect(defaultProps.onExportCSV).toHaveBeenCalled();
+  });
+});